requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
filelock==3.6.0
huggingface-hub==0.4.0
idna==3.3
joblib==1.1.0
lxml==4.8.0
nltk==3.7
numpy==1.22.3
openpyxl==3.0.9
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
sacremoses==0.0.49
scikit-learn==1.0.2
scipy==1.8.0
sentence-transformers==2.2.0
sentencepiece==0.1.96
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
threadpoolctl==3.1.0
tokenizers==0.11.6
torch==1.11.0
torchvision==0.12.0
tqdm==4.63.0
transformers==4.17.0
typing-extensions==4.1.1
urllib3==1.26.8
